[backend] post rebase fix
