+++ opencti-platform/opencti-graphql/src/database/redis.ts
@@ -482,405 +482,6 @@ export const lockResource = async (resources: Array<string>, opts: LockOptions =
 };
 // endregion
 
-// region opencti data stream
-const streamTrimming = conf.get('redis:trimming') || 0;
-const mapJSToStream = (event: any) => {
-  const cmdArgs: Array<string> = [];
-  Object.keys(event).forEach((key) => {
-    const value = event[key];
-    if (value !== undefined) {
-      cmdArgs.push(key);
-      cmdArgs.push(JSON.stringify(value));
-    }
-  });
-  return cmdArgs;
-};
-const pushToStream = async (context: AuthContext, user: AuthUser, client: Cluster | Redis, event: BaseEvent, opts: EventOpts = {}) => {
-  const draftContext = getDraftContext(context, user);
-  const eventToPush = { ...event, event_id: context.eventId };
-  if (!draftContext && isStreamPublishable(opts)) {
-    const pushToStreamFn = async () => {
-      if (streamTrimming) {
-        await client.call('XADD', REDIS_STREAM_NAME, 'MAXLEN', '~', streamTrimming, '*', ...mapJSToStream(eventToPush));
-      } else {
-        await client.call('XADD', REDIS_STREAM_NAME, '*', ...mapJSToStream(eventToPush));
-      }
-    };
-    await telemetry(context, user, 'INSERT STREAM', {
-      [SEMATTRS_DB_NAME]: 'stream_engine',
-    }, pushToStreamFn);
-  }
-};
-
-// Merge
-const buildMergeEvent = async (user: AuthUser, previous: StoreObject, instance: StoreObject, sourceEntities: Array<StoreObject>): Promise<MergeEvent> => {
-  const message = generateMergeMessage(instance, sourceEntities);
-  const previousStix = convertStoreToStix_2_1(previous) as StixCoreObject;
-  const currentStix = convertStoreToStix_2_1(instance) as StixCoreObject;
-  return {
-    version: EVENT_CURRENT_VERSION,
-    type: EVENT_TYPE_MERGE,
-    scope: 'external',
-    message,
-    origin: user.origin,
-    data: currentStix,
-    context: {
-      patch: jsonpatch.compare(previousStix, currentStix),
-      reverse_patch: jsonpatch.compare(currentStix, previousStix),
-      sources: await asyncListTransformation<StixObject>(sourceEntities, convertStoreToStix_2_1) as StixCoreObject[],
-    },
-  };
-};
-export const storeMergeEvent = async (
-  context: AuthContext,
-  user: AuthUser,
-  initialInstance: StoreObject,
-  mergedInstance: StoreObject,
-  sourceEntities: Array<StoreObject>,
-  opts: EventOpts,
-) => {
-  try {
-    const event = await buildMergeEvent(user, initialInstance, mergedInstance, sourceEntities);
-    await pushToStream(context, user, getClientBase(), event, opts);
-    return event;
-  } catch (e) {
-    throw DatabaseError('Error in store merge event', { cause: e });
-  }
-};
-// Update
-export const buildStixUpdateEvent = (
-  user: AuthUser,
-  previousStix: StixCoreObject,
-  stix: StixCoreObject,
-  message: string,
-  changes: Change[],
-  opts: UpdateEventOpts = {},
-): UpdateEvent => {
-  // Build and send the event
-  const patch = jsonpatch.compare(previousStix, stix);
-  const previousPatch = jsonpatch.compare(stix, previousStix);
-  if (patch.length === 0 || previousPatch.length === 0) {
-    throw UnsupportedError('Update event must contains a valid previous patch');
-  }
-  if (patch.length === 1 && patch[0].path === '/modified' && !opts.allow_only_modified) {
-    throw UnsupportedError('Update event must contains more operation than just modified/updated_at value');
-  }
-  const entityType = stix.extensions[STIX_EXT_OCTI].type;
-  const scope = INTERNAL_EXPORTABLE_TYPES.includes(entityType) ? 'internal' : 'external';
-  return {
-    version: EVENT_CURRENT_VERSION,
-    type: EVENT_TYPE_UPDATE,
-    scope,
-    message,
-    origin: user.origin,
-    data: stix,
-    commit: opts.commit,
-    noHistory: opts.noHistory,
-    context: {
-      patch,
-      reverse_patch: previousPatch,
-      related_restrictions: opts.related_restrictions,
-      pir_ids: opts.pir_ids,
-      changes,
-    },
-  };
-};
-export const publishStixToStream = async (context: AuthContext, user: AuthUser, event: StreamDataEvent) => {
-  await pushToStream(context, user, getClientBase(), event);
-};
-const buildUpdateEvent = (user: AuthUser, previous: StoreObject, instance: StoreObject, message: string, changes: Change[], opts: UpdateEventOpts): UpdateEvent => {
-  // Build and send the event
-  const stix = convertStoreToStix_2_1(instance) as StixCoreObject;
-  const previousStix = convertStoreToStix_2_1(previous) as StixCoreObject;
-  return buildStixUpdateEvent(user, previousStix, stix, message, changes, opts);
-};
-export const storeUpdateEvent = async (
-  context: AuthContext,
-  user: AuthUser,
-  previous: StoreObject,
-  instance: StoreObject,
-  message: string,
-  changes: Change[],
-  opts: UpdateEventOpts = {},
-) => {
-  try {
-    if (isStixExportableInStreamData(instance)) {
-      const event = buildUpdateEvent(user, previous, instance, message, changes, opts);
-      await pushToStream(context, user, getClientBase(), event, opts);
-      return event;
-    }
-    return undefined;
-  } catch (e) {
-    throw DatabaseError('Error in store update event', { cause: e });
-  }
-};
-// Create
-export const buildCreateEvent = (user: AuthUser, instance: StoreObject, message: string): StreamDataEvent => {
-  const stix = convertStoreToStix_2_1(instance) as StixCoreObject;
-  return {
-    version: EVENT_CURRENT_VERSION,
-    type: EVENT_TYPE_CREATE,
-    scope: INTERNAL_EXPORTABLE_TYPES.includes(instance.entity_type) ? 'internal' : 'external',
-    message,
-    origin: user.origin,
-    data: stix,
-  };
-};
-export const storeCreateRelationEvent = async (context: AuthContext, user: AuthUser, instance: StoreRelation, opts: CreateEventOpts = {}) => {
-  try {
-    if (isStixExportableInStreamData(instance)) {
-      const { withoutMessage = false, restore = false } = opts;
-      let message = '-';
-      if (!withoutMessage) {
-        message = restore ? generateRestoreMessage(instance) : generateCreateMessage(instance);
-      }
-      const event = buildCreateEvent(user, instance, message);
-      await pushToStream(context, user, getClientBase(), event, opts);
-      return event;
-    }
-    return undefined;
-  } catch (e) {
-    throw DatabaseError('Error in store create relation event', { cause: e });
-  }
-};
-export const storeCreateEntityEvent = async (context: AuthContext, user: AuthUser, instance: StoreObject, message: string, opts: CreateEventOpts = {}) => {
-  try {
-    if (isStixExportableInStreamData(instance)) {
-      const event = buildCreateEvent(user, instance, message);
-      await pushToStream(context, user, getClientBase(), event, opts);
-      return event;
-    }
-    return undefined;
-  } catch (e) {
-    throw DatabaseError('Error in store create entity event', { cause: e });
-  }
-};
-
-// Delete
-export const buildDeleteEvent = async (
-  user: AuthUser,
-  instance: StoreObject,
-  message: string,
-): Promise<DeleteEvent> => {
-  const stix = convertStoreToStix_2_1(instance) as StixCoreObject;
-  return {
-    version: EVENT_CURRENT_VERSION,
-    type: EVENT_TYPE_DELETE,
-    scope: INTERNAL_EXPORTABLE_TYPES.includes(instance.entity_type) ? 'internal' : 'external',
-    message,
-    origin: user.origin,
-    data: stix,
-  };
-};
-export const storeDeleteEvent = async (context: AuthContext, user: AuthUser, instance: StoreObject, opts: EventOpts = {}) => {
-  try {
-    if (isStixExportableInStreamData(instance)) {
-      const message = generateDeleteMessage(instance);
-      const event = await buildDeleteEvent(user, instance, message);
-      await pushToStream(context, user, getClientBase(), event, opts);
-      return event;
-    }
-    return undefined;
-  } catch (e) {
-    throw DatabaseError('Error in store delete event', { cause: e });
-  }
-};
-
-const mapStreamToJS = ([id, data]: any): SseEvent<any> => {
-  const count = data.length / 2;
-  const obj: any = {};
-  for (let i = 0; i < count; i += 1) {
-    obj[data[2 * i]] = JSON.parse(data[2 * i + 1]);
-  }
-  return { id, event: obj.type, data: obj };
-};
-export const fetchStreamInfo = async (streamName = REDIS_STREAM_NAME) => {
-  const res: any = await getClientBase().xinfo('STREAM', streamName);
-  const info: any = R.fromPairs(R.splitEvery(2, res) as any);
-  const firstId = info['first-entry'][0];
-  const firstEventDate = utcDate(parseInt(firstId.split('-')[0], 10)).toISOString();
-  const lastId = info['last-entry'][0];
-  const lastEventDate = utcDate(parseInt(lastId.split('-')[0], 10)).toISOString();
-  return { lastEventId: lastId, firstEventId: firstId, firstEventDate, lastEventDate, streamSize: info.length };
-};
-
-const processStreamResult = async (results: Array<any>, callback: any, withInternal: boolean | undefined) => {
-  const transform = (r: any) => mapStreamToJS(r);
-  const filter = (s: any) => (withInternal ? true : (s.data.scope ?? 'external') === 'external');
-  const events = await asyncMap(results, transform, filter);
-  const lastEventId = events.length > 0 ? R.last(events)?.id : `${new Date().valueOf()}-0`;
-  await callback(events, lastEventId);
-  return lastEventId;
-};
-
-const STREAM_BATCH_TIME = 5000;
-const MAX_RANGE_MESSAGES = 100;
-
-export interface StreamProcessor {
-  info: () => Promise<object>;
-  start: (from: string | undefined) => Promise<void>;
-  shutdown: () => Promise<void>;
-  running: () => boolean;
-}
-
-interface StreamOption {
-  withInternal?: boolean;
-  bufferTime?: number;
-  autoReconnect?: boolean;
-  streamName?: string;
-  streamBatchSize?: number;
-}
-
-export const createStreamProcessor = <T extends BaseEvent> (
-  _user: AuthUser,
-  provider: string,
-  callback: (events: Array<SseEvent<T>>, lastEventId: string) => void,
-  opts: StreamOption = {},
-): StreamProcessor => {
-  let client: Cluster | Redis;
-  let startEventId: string;
-  let processingLoopPromise: Promise<void>;
-  let streamListening = true;
-  const streamName = opts.streamName ?? REDIS_STREAM_NAME;
-
-  const processStep = async () => {
-    // since previous call is async (and blocking) we should check if we are still running before processing the message
-    if (!streamListening) {
-      return false;
-    }
-    try {
-      // Consume the data stream
-      const streamResult = await client.call(
-        'XREAD',
-        'COUNT',
-        MAX_RANGE_MESSAGES,
-        'BLOCK',
-        STREAM_BATCH_TIME,
-        'STREAMS',
-        streamName,
-        startEventId,
-      ) as any[];
-      // Process the event results
-      if (streamResult && streamResult.length > 0) {
-        const [, results] = streamResult[0];
-        const lastElementId = await processStreamResult(results, callback, opts.withInternal);
-        startEventId = lastElementId || startEventId;
-      } else {
-        await processStreamResult([], callback, opts.withInternal);
-      }
-      await wait(opts.bufferTime ?? 50);
-    } catch (err) {
-      logApp.error('Redis stream consume fail', { cause: err, provider });
-      if (opts.autoReconnect) {
-        await waitInSec(5);
-      } else {
-        return false;
-      }
-    }
-    return streamListening;
-  };
-  const processingLoop = async () => {
-    while (streamListening) {
-      if (!(await processStep())) {
-        streamListening = false;
-        break;
-      }
-    }
-  };
-  return {
-    info: async () => fetchStreamInfo(streamName),
-    running: () => streamListening,
-    start: async (start = 'live') => {
-      if (streamListening) {
-        let fromStart = start;
-        if (isEmptyField(fromStart)) {
-          fromStart = 'live';
-        }
-        startEventId = fromStart === 'live' ? '$' : fromStart;
-        logApp.info('[STREAM] Starting stream processor', { provider, startEventId });
-        processingLoopPromise = (async () => {
-          client = await createRedisClient(provider, opts.autoReconnect); // Create client for this processing loop
-          try {
-            await processingLoop();
-          } finally {
-            logApp.info('[STREAM] Stream processor terminated, closing Redis client');
-            client.disconnect();
-          }
-        })();
-      }
-    },
-    shutdown: async () => {
-      logApp.info('[STREAM] Shutdown stream processor', { provider });
-      streamListening = false;
-      if (processingLoopPromise) {
-        await processingLoopPromise;
-      }
-      logApp.info('[STREAM] Stream processor current promise terminated');
-    },
-  };
-};
-// endregion
-
-// region fetch stream event range
-export const fetchStreamEventsRangeFromEventId = async (
-  client: Cluster | Redis,
-  startEventId: string,
-  callback: (events: Array<SseEvent<DataEvent>>, lastEventId: string) => void,
-  opts: StreamOption = {},
-) => {
-  const { streamBatchSize = MAX_RANGE_MESSAGES } = opts;
-  let effectiveStartEventId = startEventId;
-  try {
-    // Consume streamBatchSize number of stream events from startEventId (excluded)
-    const streamResult = await client.call(
-      'XRANGE',
-      opts.streamName ?? REDIS_STREAM_NAME,
-      `(${startEventId}`, // ( prefix to exclude startEventId
-      '+',
-      'COUNT',
-      streamBatchSize,
-    ) as any[];
-    // Process the event results
-    if (streamResult && streamResult.length > 0) {
-      const lastStreamResultId = R.last(streamResult)[0]; // id of last event fetched (internal or external)
-      await processStreamResult(streamResult, callback, opts.withInternal); // process the stream events of the range
-      if (lastStreamResultId) {
-        effectiveStartEventId = lastStreamResultId;
-      }
-    } else {
-      await processStreamResult([], callback, opts.withInternal);
-    }
-  } catch (err) {
-    logApp.error('Redis stream consume fail', { cause: err });
-    if (opts.autoReconnect) {
-      await waitInSec(2);
-    }
-  }
-  return { lastEventId: effectiveStartEventId };
-};
-
-// region opencti notification stream
-export const NOTIFICATION_STREAM_NAME = `${REDIS_PREFIX}stream.notification`;
-const notificationTrimming = conf.get('redis:notification_trimming') || 50000;
-export const storeNotificationEvent = async (context: AuthContext, event: any) => {
-  await getClientBase().call('XADD', NOTIFICATION_STREAM_NAME, 'MAXLEN', '~', notificationTrimming, '*', ...mapJSToStream(event));
-};
-export const fetchRangeNotifications = async <T extends BaseEvent> (start: Date, end: Date): Promise<Array<T>> => {
-  const streamResult = await getClientBase().call('XRANGE', NOTIFICATION_STREAM_NAME, start.getTime(), end.getTime()) as any[];
-  const streamElements: Array<SseEvent<T>> = R.map((r) => mapStreamToJS(r), streamResult);
-  return streamElements.filter((s) => s.event === 'live').map((e) => e.data);
-};
-// endregion
-
-// region opencti audit stream
-export const EVENT_ACTIVITY_VERSION = '1';
-export const ACTIVITY_STREAM_NAME = `${REDIS_PREFIX}stream.activity`;
-const auditTrimming = conf.get('redis:activity_trimming') || 50000;
-export const storeActivityEvent = async (event: ActivityStreamEvent) => {
-  await getClientBase().call('XADD', ACTIVITY_STREAM_NAME, 'MAXLEN', '~', auditTrimming, '*', ...mapJSToStream(event));
-};
-// endregion
-
 // region work handling
 export const redisDeleteWorks = async (internalIds: Array<string>) => {
   const ids = Array.isArray(internalIds) ? internalIds : [internalIds];
